#184 #187
*. And some major refactoring.

+++ lib-artnet/src/artnetnodegetportaddress.cpp
@@ -34,7 +34,7 @@
 
 using namespace artnet;
 
-bool ArtNetNode::GetPortAddress(uint32_t nPortIndex, uint16_t &nAddress, PortDir dir) const {
+bool ArtNetNode::GetPortAddress(uint8_t nPortIndex, uint16_t &nAddress, PortDir dir) const {
 	if (dir == PortDir::INPUT) {
 		assert(nPortIndex < ARTNET_NODE_MAX_PORTS_INPUT);
