+++ lib-artnet4/include/artnet4node.h
@@ -37,12 +37,12 @@
 
 class ArtNet4Node: public ArtNetNode, public ArtNet4Handler {
 public:
-	ArtNet4Node(uint32_t nPages = 1);
+	ArtNet4Node(uint8_t nPages = 1);
 
-	void SetPort(uint32_t nPortIndex, artnet::PortDir dir) override;
+	void SetPort(uint8_t nPortIndex, artnet::PortDir dir) override;
 
 	void HandleAddress(uint8_t nCommand) override;
-	uint8_t GetStatus(uint32_t nPortIndex) override;
+	uint8_t GetStatus(uint8_t nPortIndex) override;
 
 	bool IsStatusChanged() override {
 		return m_Bridge.IsStatusChanged();
