+++ lib-dmx/include/dmxset.h
@@ -42,7 +42,7 @@ public:
 
 	virtual void SetPortDirection(uint32_t nPort, dmx::PortDirection tPortDirection, bool bEnableData)=0;
 
-	virtual void RdmSendRaw(uint32_t nPort, const uint8_t *pRdmData, uint16_t nLength)=0;
+	virtual void RdmSendRaw(uint32_t nPort, const uint8_t *pRdmData, uint32_t nLength)=0;
 
 	virtual const uint8_t *RdmReceive(uint32_t nPort)=0;
 	virtual const uint8_t *RdmReceiveTimeOut(uint32_t nPort, uint32_t nTimeOut)=0;
