+++ lib-dmxreceiver/include/dmxreceiver.h
@@ -50,13 +50,13 @@ public:
 	void Print() {}
 
 private:
-	bool IsDmxDataChanged(const uint8_t *pData, uint16_t nLength);
+	bool IsDmxDataChanged(const uint8_t *pData, uint32_t nLength);
 
 private:
 	LightSet *m_pLightSet { nullptr };
 	bool m_IsActive { false };
 	uint8_t m_Data[dmx::buffer::SIZE]; // With DMX Start Code
-	uint16_t m_nLength { 0 };
+	uint32_t m_nLength { 0 };
 };
 
 #endif /* DMXCONTROLLER_H_ */
