+++ lib-esp8266/include/esp8266.h
@@ -41,14 +41,14 @@ extern void esp8266_write_4bits(const uint8_t);
 extern void esp8266_write_byte(const uint8_t);
 extern void esp8266_write_halfword(const uint16_t);
 extern void esp8266_write_word(const uint32_t);
-extern void esp8266_write_bytes(const uint8_t *, const uint16_t);
+extern void esp8266_write_bytes(const uint8_t *, const uint32_t);
 extern void esp8266_write_str(const char *);
 
 extern uint8_t esp8266_read_byte(void);
 extern void esp8266_read_bytes(/*@out@*/const uint8_t *, const uint16_t);
 extern uint16_t esp8266_read_halfword(void);
 extern uint32_t esp8266_read_word(void);
-extern void esp8266_read_str(/*@out@*/char *, uint16_t *);
+extern void esp8266_read_str(/*@out@*/char *, uint32_t *);
 
 #ifdef __cplusplus
 }
