+++ lib-ltc/include/h3/systimereader.h
@@ -42,7 +42,7 @@ public:
 	void Print();
 
 	//
-	void HandleRequest(void *pBuffer = nullptr, uint32_t nBufferLength = 0);
+	void HandleRequest(void *pBuffer = nullptr, uint16_t nBufferLength = 0);
 
 	// Control
 	void ActionStart();
